Daily20140327_2

+++ src/main/java/org/minnie/utility/util/FileUtil.java
@@ -67,6 +67,11 @@ public class FileUtil {
 	 */
 	public static boolean deleteFile(String filePath) {
 		File file = new File(filePath);
+		
+		if(!file.exists()){
+			return true;
+		}
+		
 		return deleteFile(file);
 	}
